+++ ProcessMaker/Providers/AuthServiceProvider.php
@@ -67,31 +67,9 @@ class AuthServiceProvider extends ServiceProvider
         });
 
         try {
-            $permissions = Permission::select('name')->get();
-            // Define the Gate permissions
-            $permissions->each(function ($permission) {
-                Gate::define($permission->name, function (User $user, ...$params) use ($permission) {
-                    $authorized = false;
-
-                    // Check if the user has the permission
-                    if ($user->hasPermission($permission->name)) {
-                        return true;
-                    }
-
-                    // If the user has no projects, return false.
-                    $projects = $this->getProjectsForUser($user->id);
-                    if (empty($projects)) {
-                        return false;
-                    }
-
-                    // Check if the user has 'create-projects' permission and the request is from specific endpoints
-                    // Users that ONLY have 'create-projects' permission are allowed to access specific endpoints
-                    $isAllowedEndpoint = $this->checkAllowedEndpoints($projects, request()->path());
-                    if ($user->hasPermission('create-projects') && $isAllowedEndpoint) {
-                        $authorized = $this->isProjectAsset($permission, $params);
-                    }
-
-                    return $authorized;
+            Permission::select('name')->get()->each(function ($permission) {
+                Gate::define($permission->name, function ($user) use ($permission) {
+                    return $user->hasPermission($permission->name);
                 });
             });
         } catch (\Exception $e) {
@@ -105,144 +83,5 @@ class AuthServiceProvider extends ServiceProvider
 
             return app(AnonymousUser::class);
         });
-    }
-
-    /**
-     * Retrieves the projects associated with the user.
-     */
-    private function getProjectsForUser(int $userId): array
-    {
-        if (!hasPackage('package-projects')) {
-            return [];
-        }
-
-        // Get projects where the current user is an owner
-        $ownerProjects = DB::table('projects')
-            ->where('user_id', $userId)
-            ->pluck('id');
-
-        // Get projects where the current user is a member
-        $memberProjects = DB::table('projects')
-            ->join('project_members', 'projects.id', '=', 'project_members.project_id')
-            ->where([
-                'project_members.member_id' => $userId,
-                'project_members.member_type' => User::class,
-            ])
-            ->pluck('projects.id');
-
-        // Combine owner and member projects
-        return $ownerProjects->merge($memberProjects)->unique()->values()->toArray();
-    }
-
-    /**
-     * Checks if the current path is allowed based on the user's project assets.
-     */
-    private function checkAllowedEndpoints(array $projectIds, string $currentPath): bool
-    {
-        $allowedEndpoints = [
-            'api',
-        ];
-
-        $dataSourceClass = 'ProcessMaker\Packages\Connectors\DataSources\Models\DataSource';
-        $decisionTableClass = 'ProcessMaker\Package\PackageDecisionEngine\Models\DecisionTable';
-        $flowGenieTableClass = 'ProcessMaker\Package\PackageAi\Models\FlowGenie';
-
-        // Get the assets associated with the user's projects
-        $projectAssets = DB::table('project_assets')
-            ->select('asset_id', 'asset_type')
-            ->whereIn('project_id', $projectIds)
-            ->distinct()
-            ->get();
-
-        foreach ($projectAssets as $asset) {
-            // Get each project asset's type and id
-            $assetId = $asset->asset_id;
-            $assetType = $asset->asset_type;
-
-            // Check asset types and push to $allowedEndpoints
-            if ($assetType === Process::class) {
-                $allowedEndpoints[] = "modeler/{$assetId}";
-            } elseif ($assetType === Screen::class) {
-                $allowedEndpoints[] = "designer/screen-builder/{$assetId}/edit";
-                $allowedEndpoints[] = "designer/screens/{$assetId}/edit";
-                $allowedEndpoints[] = 'designer/screens/preview';
-            } elseif ($assetType === Script::class) {
-                $allowedEndpoints[] = "designer/scripts/{$assetId}/builder";
-                $allowedEndpoints[] = "designer/scripts/{$assetId}/edit";
-                $allowedEndpoints[] = 'designer/scripts/preview';
-            }
-
-            if (class_exists($dataSourceClass) && $assetType === $dataSourceClass) {
-                $allowedEndpoints[] = "designer/data-sources/{$assetId}/edit";
-            }
-            if (class_exists($decisionTableClass) && $assetType === $decisionTableClass) {
-                $allowedEndpoints[] = "decision-tables/table-builder/{$assetId}/edit";
-            }
-            if (class_exists($flowGenieTableClass) && $assetType === $flowGenieTableClass) {
-                $allowedEndpoints[] = "designer/flow-genies/{$assetId}/edit";
-            }
-        }
-
-        return Str::contains($currentPath, $allowedEndpoints);
-    }
-
-    private function isProjectAsset($permission, $params)
-    {
-        if ($params && $params[0]) {
-            return $this->handleUpdateDeleteOperations($permission, class_basename($params[0]));
-        }
-
-        return $this->checkForListCreateOperations($permission);
-    }
-
-    private function handleUpdateDeleteOperations($permission, $modelClass)
-    {
-        $asset = $this->getAssetName($modelClass);
-
-        return $this->checkPermissionForAsset($permission, $asset);
-    }
-
-    /**
-     * Get the asset name based on the model class.
-     *
-     * @param string $modelClass
-     * @return string
-     */
-    private function getAssetName($modelClass)
-    {
-        $asset = Str::snake(class_basename($modelClass));
-
-        // Adjust asset name for DataSource class
-        if ($modelClass === 'DataSource') {
-            $asset = 'data-source';
-        }
-
-        return $asset;
-    }
-
-    private function checkForListCreateOperations($permission)
-    {
-        $projectAssetTypes = [
-            'process',
-            'screen',
-            'script',
-            'data-source',
-            'decision_table',
-            'pm-block',
-            'flow_genies'
-        ];
-
-        foreach ($projectAssetTypes as $asset) {
-            if (Str::contains($permission->name, $asset)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function checkPermissionForAsset($permission, $asset)
-    {
-        return Str::contains($permission->name, $asset);
     }
 }
